Implement Sort command
Let's allow the user to sort their applications by date (following the
order laid out in the DG).

Bulk of changes to logic lie in `UniqueInternshipList.java`.
We only perform mutations on `baseInternalList` and maintain a sorted
wrapper `internalListSortedByDate` on this.

Sorting is then a matter of populating the `uiFacingInternalList` with
the correct underlying list elements.

In other words, `baseInternalList` and `internalListSortedByDate` should
always have the same elements but the former is in the order that
applications were added, while the latter follows the date ordering laid
out in the DG.

As always, feedback on design improvements are welcome and appreciated.
Finally, note that the current Sort command does not interact well with Undo.
List is used to reset both ordering and filtering (see
`ListCommand.java`)